# Improve perf of CredentialCache.GetCredential