@rzikm 2024-06-19
/benchmark micro aspnet-perf-lin libs --variable filter=GetCredential_Uri