@MihaZupan 2024-06-19
@EgorBot
```c#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using System;
using System.Collections.Generic;
using System.Net;

public class CredentialCacheTests
{
    private const string UriPrefix = "http://name";
    private const string HostPrefix = "name";
    private const int Port = 80;
    private const string AuthenticationType = "authType";

    private static readonly NetworkCredential s_credential = new NetworkCredential();

    private readonly Dictionary<(int uriCount, int hostPortCount), CredentialCache> _caches = new()
    {
        { (0, 0), CreateCredentialCache(0, 0) },
        { (0, 10), CreateCredentialCache(0, 10) },
        { (10, 0), CreateCredentialCache(10, 0) },
        { (10, 10), CreateCredentialCache(10, 10) }
    };

    [Benchmark]
    [Arguments("http://notfound", 0)]
    [Arguments("http://notfound", 10)]
    [Arguments("http://name5", 10)]
    public NetworkCredential GetCredential_Uri(string uriString, int uriCount)
    {
        CredentialCache cc = _caches[(uriCount: uriCount, hostPortCount: 0)];

        return cc.GetCredential(new Uri(uriString), AuthenticationType);
    }

    private static CredentialCache CreateCredentialCache(int uriCount, int hostPortCount)
    {
        var cc = new CredentialCache();

        for (int i = 0; i < uriCount; i++)
        {
            Uri uri = new Uri(UriPrefix + i.ToString());
            cc.Add(uri, AuthenticationType, s_credential);
        }

        for (int i = 0; i < hostPortCount; i++)
        {
            string host = HostPrefix + i.ToString();
            cc.Add(host, Port, AuthenticationType, s_credential);
        }

        return cc;
    }
}
```