@rzikm 2024-06-24
I took a look at why we regressed the one scenario, and it seems that that particular benchmark is measuring a corner case. I locally editted the benchmarks so that the creds in the cache on the same host, but different prefix (`http://name/prefix{i}/`) and added couple more test cases and the changes actually improve common scenario. Numbers below are this PR compared to main before introducing the regression.

```
BenchmarkDotNet v0.13.13-nightly.20240311.145, Windows 11 (10.0.22631.3737/23H2/2023Update/SunValley3)
Intel Core i9-10900K CPU 3.70GHz, 1 CPU, 20 logical and 10 physical cores
.NET SDK 9.0.100-preview.5.24307.3
  [Host]     : .NET 9.0.0 (9.0.24.30607), X64 RyuJIT AVX2
  Job-WTZYSK : .NET 9.0.0 (42.42.42.42424), X64 RyuJIT AVX2
  Job-CDIHKU : .NET 9.0.0 (42.42.42.42424), X64 RyuJIT AVX2

PowerPlanMode=00000000-0000-0000-0000-000000000000  IterationTime=250ms  MaxIterationCount=20
MinIterationCount=15  WarmupCount=1

| Method            | Job        | Toolchain          | uriString                        | uriCount | Mean      | Error    | StdDev   | Median    | Min       | Max       | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|------------------ |----------- |------------------- |--------------------------------- |--------- |----------:|---------:|---------:|----------:|----------:|----------:|------:|--------:|-------:|----------:|------------:|
| GetCredential_Uri | Job-WTZYSK | \9.0.0\corerun.exe | http://differentHost             | 10       | 273.23 ns | 4.014 ns | 3.755 ns | 273.96 ns | 265.50 ns | 276.82 ns |  1.21 |    0.02 | 0.0248 |     264 B |        1.43 |
| GetCredential_Uri | Job-CDIHKU | \main\corerun.exe  | http://differentHost             | 10       | 225.99 ns | 3.684 ns | 3.266 ns | 225.76 ns | 221.59 ns | 232.48 ns |  1.00 |    0.02 | 0.0171 |     184 B |        1.00 |
|                   |            |                    |                                  |          |           |          |          |           |           |           |       |         |        |           |             |
| GetCredential_Uri | Job-WTZYSK | \9.0.0\corerun.exe | http://name/diff/path            | 10       | 317.86 ns | 5.780 ns | 5.124 ns | 317.80 ns | 308.61 ns | 326.92 ns |  0.90 |    0.02 | 0.0259 |     272 B |        1.00 |
| GetCredential_Uri | Job-CDIHKU | \main\corerun.exe  | http://name/diff/path            | 10       | 352.02 ns | 6.018 ns | 5.629 ns | 349.61 ns | 346.69 ns | 365.42 ns |  1.00 |    0.02 | 0.0258 |     272 B |        1.00 |
|                   |            |                    |                                  |          |           |          |          |           |           |           |       |         |        |           |             |
| GetCredential_Uri | Job-WTZYSK | \9.0.0\corerun.exe | http://name/differentPrefix/path | 10       | 456.69 ns | 8.940 ns | 8.363 ns | 454.95 ns | 445.83 ns | 470.55 ns |  0.99 |    0.02 | 0.0270 |     288 B |        1.00 |
| GetCredential_Uri | Job-CDIHKU | \main\corerun.exe  | http://name/differentPrefix/path | 10       | 459.11 ns | 7.361 ns | 6.885 ns | 457.80 ns | 448.95 ns | 474.82 ns |  1.00 |    0.02 | 0.0263 |     288 B |        1.00 |
|                   |            |                    |                                  |          |           |          |          |           |           |           |       |         |        |           |             |
| GetCredential_Uri | Job-WTZYSK | \9.0.0\corerun.exe | http://name/prefix5/path         | 10       | 381.24 ns | 7.080 ns | 6.623 ns | 380.00 ns | 372.30 ns | 392.75 ns |  0.81 |    0.02 | 0.0258 |     272 B |        1.00 |
| GetCredential_Uri | Job-CDIHKU | \main\corerun.exe  | http://name/prefix5/path         | 10       | 469.81 ns | 5.792 ns | 5.418 ns | 469.72 ns | 460.92 ns | 478.05 ns |  1.00 |    0.02 | 0.0246 |     272 B |        1.00 |
|                   |            |                    |                                  |          |           |          |          |           |           |           |       |         |        |           |             |
| GetCredential_Uri | Job-WTZYSK | \9.0.0\corerun.exe | http://notfound                  | 0        |  59.25 ns | 1.229 ns | 1.150 ns |  58.77 ns |  57.81 ns |  61.34 ns |  0.98 |    0.02 | 0.0051 |      56 B |        1.00 |
| GetCredential_Uri | Job-CDIHKU | \main\corerun.exe  | http://notfound                  | 0        |  60.65 ns | 0.773 ns | 0.723 ns |  60.47 ns |  59.68 ns |  62.08 ns |  1.00 |    0.02 | 0.0052 |      56 B |        1.00 |
```